Moved all classes into a classes folder
